requirements.txt
openai
datasets
python-dotenv
tqdm

